[ROSAPPS] Add vgafontedit to build. Convert the language resource files into UTF-8. Fix x64 warnings. Patch by Hermes Belusca. Bug 7152.
svn path=/trunk/; revision=57200

+++ rosapps/applications/devutils/vgafontedit/editglyphdlg.c
@@ -40,7 +40,7 @@ InitToolbox(IN PEDIT_GLYPH_INFO Info)
     // Add the Toolbox bitmaps
     tbab.hInst = hInstance;
     tbab.nID = IDB_EDIT_GLYPH_TOOLBOX;
-    iBitmap = SendMessageW(hToolbar, TB_ADDBITMAP, 0, (LPARAM)&tbab);
+    iBitmap = (INT)SendMessageW(hToolbar, TB_ADDBITMAP, 0, (LPARAM)&tbab);
 
     AddToolboxButton(hToolbar, iBitmap + TOOLBOX_PEN, ID_TOOLBOX_PEN, TBSTATE_ENABLED | TBSTATE_CHECKED);
 }
